refactor: split man-page constants
