+++ src/constants.mjs
@@ -10,13 +10,3 @@ export const DOC_NODE_CHANGELOG_URL =
 // This is the perma-link within the API docs that reference the Stability Index
 export const DOC_API_STABILITY_SECTION_REF_URL =
   'documentation.html#stability-index';
-
-// https://github.com/nodejs/node/blob/main/doc/api/cli.md#options
-// This slug should reference the section where the available
-// options are defined.
-export const DOC_SLUG_OPTIONS = 'options';
-
-// https://github.com/nodejs/node/blob/main/doc/api/cli.md#environment-variables-1
-// This slug should reference the section where the available
-// environment variables are defined.
-export const DOC_SLUG_ENVIRONMENT = 'environment-variables-1';
