+++ src/generators/man-page/constants.mjs
@@ -1,0 +1,11 @@
+'use strict';
+
+// https://github.com/nodejs/node/blob/main/doc/api/cli.md#options
+// This slug should reference the section where the available
+// options are defined.
+export const DOC_SLUG_OPTIONS = 'options';
+
+// https://github.com/nodejs/node/blob/main/doc/api/cli.md#environment-variables-1
+// This slug should reference the section where the available
+// environment variables are defined.
+export const DOC_SLUG_ENVIRONMENT = 'environment-variables-1';
